feat: check lab6 part 1
