+++ TH/lab 6a - Building Question-Answering with watsonx.ai and Streamlit with Retrieval Augmented Generation (Transient)/Level_3/app.py
@@ -15,24 +15,12 @@ from langchain.embeddings import (HuggingFaceHubEmbeddings,
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain.vectorstores import FAISS, Chroma
 from PIL import Image
-from sentence_transformers import SentenceTransformer, models
+
 from langChainInterface import LangChainInterface
 from langchain.embeddings.openai import OpenAIEmbeddings
 import numpy as np
 import faiss
-import re
-
-def filter_text(input_text):
-    # Regular expression to match Thai and English characters and numbers
-    pattern = re.compile(r'[0-9A-Za-zก-๙\s]')
-    
-    # Use findall to extract all matching characters
-    filtered_characters = re.findall(pattern, input_text)
-    
-    # Join the characters back into a string
-    filtered_text = ''.join(filtered_characters)
-
-    return filtered_text
+from function import vector_search, get_model, generate_prompt, format_docs_for_display, format_documents
 
 # Most GENAI logs are at Debug level.
 logging.basicConfig(level=os.environ.get("LOGLEVEL", "DEBUG"))
@@ -108,57 +96,7 @@ def read_pdf(uploaded_files,chunk_size =1000,chunk_overlap=20):
              docs = text_splitter.split_documents(data)
              return docs
 
-def vector_search(query, model, index, num_results=10):
-    """
-    Transforms query to vector using a pretrained model and finds similar vectors using FAISS.
-    
-    Args:
-        query (str): User query that should be more than a sentence long.
-        model: Model used for generating embeddings.
-        index (faiss.IndexIDMap): FAISS index.
-        num_results (int): Number of results to return.
-    
-    Returns:
-        D (:obj:`numpy.array` of `float`): Distance between results and query.
-        I (:obj:`numpy.array` of `int`): Document IDs of the results.
-    """
-    vector = model.encode([query])
-    D, I = index.search(np.array(vector).astype("float32"), k=num_results)
-    return D, I
-
-
-def get_model(model_name='airesearch/wangchanberta-base-att-spm-uncased', max_seq_length=768, condition=True):
-    if condition:
-        # model_name = 'airesearch/wangchanberta-base-att-spm-uncased'
-        # model_name = "hkunlp/instructor-large"
-        word_embedding_model = models.Transformer(model_name, max_seq_length=max_seq_length)
-        pooling_model = models.Pooling(word_embedding_model.get_word_embedding_dimension(),pooling_mode='cls') # We use a [CLS] token as representation
-        model = SentenceTransformer(modules=[word_embedding_model, pooling_model])
-    return model
-
-def generate_prompt(question, context, model_type="llama-2"):
-    if model_type =="llama-2":
-        output = f"""[INST] <<SYS>>
-You are a helpful, respectful Thai assistant. Always answer as helpfully as possible, while being safe. Your answers should not include any harmful, unethical, racist, sexist, toxic, dangerous, or illegal content. Please ensure that your responses are socially unbiased and positive in nature. If a question does not make any sense, or is not factually coherent, explain why instead of answering something not correct. If you don't know the answer to a question, please don't share false information.
-
-You will receive HR Policy on user queries HR POLICY DETAILS, and QUESTION from user in the ''' below. Answer the question in Thai.
-'''
-HR POLICY DETAILS:
-{context}
-
-QUESTION: {question}
-'''
-Answer the QUESTION use details about HR Policy from HR POLICY DETAILS, explain your reasonings if the question is not related to REFERENCE please Answer
-“I don’t know the answer, it is not part of the provided HR Policy”.
-<</SYS>>
-
-คำถาม: {question} [/INST]
-คำตอบของถามเป็นภาษาไทย:
-    """
-    else:
-        return "Only llama-2 format at the moment, fill here to add other prompt templates for other model types."
-    return output
-
+# embeddings_model = get_model(model_name='airesearch/wangchanberta-base-att-spm-uncased', max_seq_length=768)
 embeddings_model = get_model(model_name='kornwtp/simcse-model-phayathaibert', max_seq_length=768)
 @st.cache_data
 def read_push_embeddings():
@@ -197,25 +135,6 @@ def read_push_embeddings():
 
     return index
 
-def format_documents(docs):
-    formatted_docs = []
-    for i, doc in enumerate(docs):
-        formatted_doc = {
-            "reference_number": i + 1,  # Assuming you want to start numbering from 1
-            "page_content": doc.page_content,
-            "source": "page " + str(doc.metadata.get('page', 'Unknown'))
-        }
-        formatted_docs.append(formatted_doc)
-    return formatted_docs
-
-def format_docs_for_display(docs):
-    formatted_text = ""
-    for doc in docs:
-        formatted_text += f"Reference Number: {doc['reference_number']}\n"
-        formatted_text += f"Source: {doc['source']}\n"
-        formatted_text += f"Content:\n{doc['page_content']}\n"
-        formatted_text += "-" * 40 + "\n"  # Separator
-    return formatted_text
 
 
 
@@ -246,6 +165,6 @@ if user_question := st.text_input(
     # Call the function with your formatted_docs
     formatted_text_for_display = format_docs_for_display(formated_search)
 
-    st.text_area(label="Model Response", value=filter_text(response), height=300)
+    st.text_area(label="Model Response", value=response, height=300)
     st.text_area(label="Reference", value=formatted_text_for_display, height=300)
     st.write()
